+++ R/scale_process.R
@@ -45,14 +45,20 @@
 #' plan(mirai_multisession, workers = 2)
 #' ncpath <- system.file("shape/nc.shp", package = "sf")
 #' ncpoly <- sf::st_read(ncpath)
+#' ncpoly <- sf::st_transform(ncpoly, "EPSG:5070")
+#'
 #' # sf object
-#' ncpnts <-
-#'   readRDS(
-#'     system.file("extdata/nc_random_point.rds", package = "chopin")
-#'   )
+#' ncpnts <- sf::st_sample(ncpoly, 2000)
+#' ncpnts <- sf::st_as_sf(ncpnts)
+#' ncpnts$pid <- seq_len(nrow(ncpnts))
+#'
 #' # file path
-#' ncelev <-
-#'     system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+#' rrast <- terra::rast(ncpoly, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
+#'
+#' # Using raster path
+#' rastpath <- file.path(tempdir(), "ncelev.tif")
+#' terra::writeRaster(rrast, rastpath, overwrite = TRUE)
 #'
 # generate grids
 #' nccompreg <-
@@ -67,12 +73,14 @@
 #'   par_grid(
 #'     grids = nccompreg,
 #'     fun_dist = extract_at,
-#'     x = ncelev,
+#'     x = rastpath,
 #'     y = ncpnts,
 #'     qsegs = 90L,
 #'     radius = 5e3L,
 #'     id = "pid"
 #'   )
+#' future::plan(future::sequential)
+#' mirai::daemons(0)
 #' @seealso
 #'  [`future::multisession`], [`future::multicore`], [`future::cluster`],
 #'  [`future.mirai::mirai_multisession`], [`future::plan`], [`par_convert_f`]
@@ -306,11 +314,22 @@ par_grid <-
 #' options(sf_use_s2 = FALSE)
 #' future::plan(future.mirai::mirai_multisession, workers = 2)
 #'
-#' ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-#' nccnty <- sf::st_read(ncpath, layer = "county")
-#' nctrct <- sf::st_read(ncpath, layer = "tracts")
-#' ncelev <-
-#'   system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+#' nccnty <- sf::st_read(
+#'   system.file("shape/nc.shp", package = "sf")
+#' )
+#' nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+#'
+#' nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+#' nccntygrid <- sf::st_as_sf(nccntygrid)
+#' nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+#' nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+#'
+#' rrast <- terra::rast(nccnty, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
+#'
+#' # Using raster path
+#' rastpath <- file.path(tempdir(), "ncelev.tif")
+#' terra::writeRaster(rrast, rastpath, overwrite = TRUE)
 #'
 #' ncsamp <-
 #'   sf::st_sample(
@@ -324,13 +343,15 @@ par_grid <-
 #' res <-
 #'   par_hierarchy(
 #'     regions = nccnty,
-#'     regions_id = "GEOID",
+#'     regions_id = "FIPS",
 #'     fun_dist = extract_at,
-#'     y = nctrct,
-#'     x = ncelev,
+#'     y = nccntygrid,
+#'     x = rastpath,
 #'     id = "GEOID",
 #'     func = "mean"
 #'   )
+#' future::plan(future::sequential)
+#' mirai::daemons(0)
 #' @importFrom future.apply future_lapply
 #' @importFrom rlang inject !!!
 #' @importFrom collapse rowbind
@@ -608,25 +629,37 @@ par_hierarchy <-
 #' sf::sf_use_s2(FALSE)
 #' future::plan(future.mirai::mirai_multisession, workers = 2)
 #'
-#' ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-#' nccnty <- sf::st_read(ncpath, layer = "county")
-#' ncelev <-
-#'   system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-#' ncelevras <- terra::rast(ncelev)
+#' nccnty <- sf::st_read(
+#'   system.file("shape/nc.shp", package = "sf")
+#' )
+#' nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+#'
+#' nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+#' nccntygrid <- sf::st_as_sf(nccntygrid)
+#' nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+#' nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+#'
+#' rrast <- terra::rast(nccnty, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
 #'
 #' tdir <- tempdir(check = TRUE)
-#' terra::writeRaster(ncelevras, file.path(tdir, "test1.tif"), overwrite = TRUE)
-#' terra::writeRaster(ncelevras, file.path(tdir, "test2.tif"), overwrite = TRUE)
+#' testpath1 <- file.path(tdir, "test1.tif")
+#' testpath2 <- file.path(tdir, "test2.tif")
+#' terra::writeRaster(rrast, testpath1, overwrite = TRUE)
+#' terra::writeRaster(rrast, testpath2, overwrite = TRUE)
 #' testfiles <- list.files(tdir, pattern = "tif$", full.names = TRUE)
 #'
 #' res <- par_multirasters(
 #'   filenames = testfiles,
 #'   fun_dist = extract_at,
-#'   x = ncelev,
+#'   x = testpath1,
 #'   y = nccnty,
 #'   id = "GEOID",
 #'   func = "mean"
 #' )
+#'
+#' future::plan(future::sequential)
+#' mirai::daemons(0)
 #' @importFrom future.apply future_lapply
 #' @importFrom terra rast
 #' @importFrom rlang inject !!!