+++ R/scale_process_mirai.R
@@ -43,14 +43,20 @@
 #' daemons(4, dispatcher = "process")
 #' ncpath <- system.file("shape/nc.shp", package = "sf")
 #' ncpoly <- sf::st_read(ncpath)
+#' ncpoly <- sf::st_transform(ncpoly, "EPSG:5070")
+#'
 #' # sf object
 #' ncpnts <-
-#'   readRDS(
-#'     system.file("extdata/nc_random_point.rds", package = "chopin")
-#'   )
+#'   sf::st_sample(ncpoly, 2000)
+#' ncpnts <- sf::st_as_sf(ncpnts)
+#' ncpnts$pid <- seq_len(nrow(ncpnts))
+#'
 #' # file path
-#' ncelev <-
-#'     system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+#' rrast <- terra::rast(ncpoly, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
+#' # Using raster path
+#' rastpath <- file.path(tempdir(), "ncelev.tif")
+#' terra::writeRaster(rrast, rastpath, overwrite = TRUE)
 #'
 # generate grids
 #' nccompreg <-
@@ -65,7 +71,7 @@
 #'   par_grid_mirai(
 #'     grids = nccompreg,
 #'     fun_dist = extract_at,
-#'     x = ncelev,
+#'     x = rastpath,
 #'     y = ncpnts,
 #'     qsegs = 90L,
 #'     radius = 5e3L,
@@ -329,11 +335,22 @@ par_grid_mirai <-
 #' options(sf_use_s2 = FALSE)
 #' mirai::daemons(4, dispatcher = "process")
 #'
-#' ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-#' nccnty <- sf::st_read(ncpath, layer = "county")
-#' nctrct <- sf::st_read(ncpath, layer = "tracts")
-#' ncelev <-
-#'   system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+#' nccnty <- sf::st_read(
+#'   system.file("shape/nc.shp", package = "sf")
+#' )
+#' nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+#'
+#' nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+#' nccntygrid <- sf::st_as_sf(nccntygrid)
+#' nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+#' nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+#'
+#' rrast <- terra::rast(nccnty, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
+#'
+#' # Using raster path
+#' rastpath <- file.path(tempdir(), "ncelev.tif")
+#' terra::writeRaster(rrast, rastpath, overwrite = TRUE)
 #'
 #' ncsamp <-
 #'   sf::st_sample(
@@ -347,10 +364,10 @@ par_grid_mirai <-
 #' res <-
 #'   par_hierarchy_mirai(
 #'     regions = nccnty,
-#'     regions_id = "GEOID",
+#'     regions_id = "FIPS",
 #'     fun_dist = extract_at,
-#'     y = nctrct,
-#'     x = ncelev,
+#'     y = nccntygrid,
+#'     x = rastpath,
 #'     id = "GEOID",
 #'     func = "mean",
 #'     .debug = TRUE
@@ -663,21 +680,28 @@ par_hierarchy_mirai <-
 #' sf::sf_use_s2(FALSE)
 #' mirai::daemons(4, dispatcher = "process")
 #'
-#' ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-#' nccnty <- sf::st_read(ncpath, layer = "county")
-#' ncelev <-
-#'   system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-#' ncelevras <- terra::rast(ncelev)
+#' nccnty <- sf::st_read(
+#'   system.file("shape/nc.shp", package = "sf")
+#' )
+#' nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+#'
+#' nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+#' nccntygrid <- sf::st_as_sf(nccntygrid)
+#' nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+#' nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+#'
+#' rrast <- terra::rast(nccnty, nrow = 600, ncol = 1320)
+#' terra::values(rrast) <- rgamma(7.92e5, 4, 2)
 #'
 #' tdir <- tempdir(check = TRUE)
-#' terra::writeRaster(ncelevras, file.path(tdir, "test1.tif"), overwrite = TRUE)
-#' terra::writeRaster(ncelevras, file.path(tdir, "test2.tif"), overwrite = TRUE)
+#' terra::writeRaster(rrast, file.path(tdir, "test1.tif"), overwrite = TRUE)
+#' terra::writeRaster(rrast, file.path(tdir, "test2.tif"), overwrite = TRUE)
 #' testfiles <- list.files(tdir, pattern = "tif$", full.names = TRUE)
 #'
 #' res <- par_multirasters_mirai(
 #'   filenames = testfiles,
 #'   fun_dist = extract_at,
-#'   x = ncelev,
+#'   x = rrast,
 #'   y = nccnty,
 #'   id = "GEOID",
 #'   func = "mean"