+++ tests/testthat/test-scale_process.R
@@ -144,12 +144,17 @@ testthat::test_that("par_grid -- plain mode with raster path", {
 
   ## Bundled random points in NC
   ncpnts <-
-    readRDS(system.file("extdata/nc_random_point.rds", package = "chopin"))
+    readRDS(rppath)
   ncpnts <- terra::vect(ncpnts)
   ncpnts <- terra::project(ncpnts, "EPSG:5070")
 
   ## Resampled SRTM data in NC
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  ras <- terra::rast(ncpoly, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
 
   ## Random points in NC
@@ -209,7 +214,7 @@ testthat::test_that("par_grid -- plain mode with raster path", {
         )
       )
   })
-
+  future::plan(future::sequential)
 })
 
 
@@ -227,19 +232,30 @@ testthat::test_that("par_grid -- grid_advanced mode", {
   )
   withr::local_seed(202407)
   # Reading data
-  ## NC counties polygon
-  ncpath <- system.file("shape/nc.shp", package = "sf")
-  ncpoly <- terra::vect(ncpath) %>%
-    terra::project("EPSG:5070")
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
 
-  ## Bundled random points in NC
-  ncpnts <-
-    readRDS(system.file("extdata/nc_random_point.rds", package = "chopin"))
-  ncpnts <- terra::vect(ncpnts)
-  ncpnts <- terra::project(ncpnts, "EPSG:5070")
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
 
   ## Resampled SRTM data in NC
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
 
   ## Random points in NC
@@ -286,7 +302,7 @@ testthat::test_that("par_grid -- grid_advanced mode", {
     "Switch terra class to sf..."
   )
 
-  testthat::expect_message(
+  testthat::expect_no_error(
     ppg2 <-
       par_pad_grid(
         input = ncsamp,
@@ -296,10 +312,7 @@ testthat::test_that("par_grid -- grid_advanced mode", {
         padding = 3e4L,
         grid_min_features = 50L,
         merge_max = 5L
-      ),
-    paste0(
-      "No grid to merge."
-    )
+      )
   )
 
   # run with grid_advanced mode grids
@@ -320,6 +333,8 @@ testthat::test_that("par_grid -- grid_advanced mode", {
       )
   })
 
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -344,14 +359,30 @@ testthat::test_that("par_grid -- grid_quantile mode", {
   ncpoly <- terra::vect(ncpath) %>%
     terra::project("EPSG:5070")
 
-  ## Bundled random points in NC
-  ncpnts <-
-    readRDS(system.file("extdata/nc_random_point.rds", package = "chopin"))
-  ncpntssf <- sf::st_transform(ncpnts, "EPSG:5070")
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
 
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
 
   ## Resampled SRTM data in NC
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
 
   ## Random points in NC
@@ -387,7 +418,7 @@ testthat::test_that("par_grid -- grid_quantile mode", {
       par_grid(
         grids = ppg,
         fun_dist = extract_at,
-        y = ncpntssf,
+        y = ncp,
         x = ncelev,
         qsegs = 90L,
         radius = 5e3L,
@@ -396,6 +427,8 @@ testthat::test_that("par_grid -- grid_quantile mode", {
     )
 
   testthat::expect_s3_class(resq, "data.frame")
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -420,15 +453,33 @@ testthat::test_that("par_grid -- par_pad_balanced", {
   ncpoly <- terra::vect(ncpath) %>%
     terra::project("EPSG:5070")
 
-  ## Bundled random points in NC
-  ncpnts <-
-    readRDS(system.file("extdata/nc_random_point.rds", package = "chopin"))
-  ncpntssf <- sf::st_transform(ncpnts, "EPSG:5070")
-  ncpntst <- terra::vect(ncpntssf)
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
+
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
 
   ## Resampled SRTM data in NC
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  ras <- terra::rast(ncpoly, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
+  ncpntst <- ncpnts
+
 
   ## Random points in NC
   ncsamp <-
@@ -440,6 +491,7 @@ testthat::test_that("par_grid -- par_pad_balanced", {
     )
   ncsamp <- terra::set.crs(ncsamp, "EPSG:5070")
   ncsamp$kid <- sprintf("K-%05d", seq(1, nrow(ncsamp)))
+
 
   tdir <- tempdir()
   target_file <- "ncrandpnts.gpkg"
@@ -488,6 +540,9 @@ testthat::test_that("par_grid -- par_pad_balanced", {
       .debug = TRUE
     )
   )
+
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -512,12 +567,31 @@ testthat::test_that(
     future::plan(future.mirai::mirai_multisession, workers = 2L)
     withr::local_seed(202407)
 
-    ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-    nccnty <- sf::st_read(ncpath, layer = "county")
-    nctrct <- sf::st_read(ncpath, layer = "tracts")
-    ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    ncelev <- terra::rast(ncelevpath)
+    nccnty <- sf::st_read(
+      system.file("shape/nc.shp", package = "sf")
+    )
+    nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+    nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+    nccntygrid <- sf::st_as_sf(nccntygrid)
+    nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+    suppressWarnings(
+      nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+    )
 
+    ## Generated random points in NC
+    data("ncpoints", package = "chopin")
+    ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+    ncp$pid <- seq_len(nrow(ncp))
+    ncpnts <- terra::vect(ncp)
+
+    ## Resampled SRTM data in NC
+    ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+    terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+    # Using raster path
+    ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+    terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
+    ncelev <- terra::rast(ncelevpath)
     ncsamp <-
       terra::spatSample(
         terra::ext(ncelev),
@@ -532,10 +606,10 @@ testthat::test_that(
       residb <-
         par_hierarchy(
           regions = nccnty,
-          regions_id = "GEOID",
+          regions_id = "FIPS",
           .debug = TRUE,
           fun_dist = extract_at,
-          y = nctrct,
+          y = nccntygrid,
           x = ncelevpath,
           id = "GEOID",
           func = "mean",
@@ -551,10 +625,10 @@ testthat::test_that(
         residb2 <-
           par_hierarchy(
             regions = nccnty,
-            regions_id = "GEOID",
+            regions_id = "FIPS",
             .debug = TRUE,
             fun_dist = extract_at,
-            y = nctrct,
+            y = nccntygrid,
             x = ncelev,
             id = "GEOID",
             func = "mean"
@@ -565,6 +639,9 @@ testthat::test_that(
              "Attempt to track the data source file path...\n")
     )
     testthat::expect_true(is.data.frame(residb2))
+
+    future::plan(future::sequential)
+    mirai::daemons(0)
 
   }
 )
@@ -588,10 +665,30 @@ testthat::test_that("par_hierarchy: multicore-SpatRaster input", {
   )
   withr::local_seed(202407)
 
-  ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-  nccnty <- sf::st_read(ncpath, layer = "county")
-  nctrct <- sf::st_read(ncpath, layer = "tracts")
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
+
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
+
+  ## Resampled SRTM data in NC
+  ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
 
   ncsamp <-
@@ -611,7 +708,7 @@ testthat::test_that("par_hierarchy: multicore-SpatRaster input", {
         regions_id = unlist(nccnty[["GEOID"]]),
         .debug = TRUE,
         fun_dist = extract_at,
-        y = nctrct,
+        y = nccntygrid,
         x = ncelev,
         id = "GEOID",
         func = "mean"
@@ -626,7 +723,7 @@ testthat::test_that("par_hierarchy: multicore-SpatRaster input", {
         regions = nccnty,
         regions_id = c(1, 2, 3),
         fun_dist = extract_at,
-        y = nctrct,
+        y = nccntygrid,
         x = ncelev,
         id = "GEOID",
         func = "mean"
@@ -634,6 +731,8 @@ testthat::test_that("par_hierarchy: multicore-SpatRaster input", {
     "The length of regions_id is not valid."
   )
 
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -653,13 +752,32 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
     )
   )
 
-  ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-  nccnty <- terra::vect(ncpath, layer = "county")
-  nctrct <- sf::st_read(ncpath, layer = "tracts")
-  nctrct <- terra::vect(nctrct)
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
+
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
+
+  ## Resampled SRTM data in NC
+  ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
-  ncroad <- system.file("extdata/ncroads_first.gpkg", package = "chopin")
+  ncroad <- terra::vect(roadpath)
   ncsamp <-
     terra::spatSample(
       terra::ext(ncelev),
@@ -670,8 +788,8 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
   ncsamp$kid <- sprintf("K-%05d", seq(1, nrow(ncsamp)))
   ncsamp <- terra::set.crs(ncsamp, "EPSG:5070")
 
-  nctrctc <- terra::centroids(nctrct)
-  ncroadv <- terra::vect(ncroad)
+  nctrctc <- terra::centroids(terra::vect(nccntygrid))
+  ncroadv <- ncroad
 
   future::plan(future::multicore(workers = 2L))
   # no errors since 100km buffer is enough to capture
@@ -679,7 +797,7 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
   resnas0 <-
     par_hierarchy(
       regions = nccnty,
-      regions_id = "GEOID",
+      regions_id = "FIPS",
       pad_y = TRUE,
       pad = 50000,
       .debug = TRUE,
@@ -695,7 +813,7 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
   resnas <-
     par_hierarchy(
       regions = nccnty,
-      regions_id = "GEOID",
+      regions_id = "FIPS",
       pad_y = TRUE,
       pad = 100000,
       .debug = TRUE,
@@ -709,12 +827,12 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
 
   testthat::skip_on_os("windows")
   # regions are sf object
-  nctrcc <- terra::centroids(nctrct)
+  nctrcc <- terra::centroids(terra::vect(nccntygrid))
   testthat::expect_no_error(
     resnasx <-
       par_hierarchy(
         regions = sf::st_as_sf(nccnty),
-        regions_id = "GEOID",
+        regions_id = "FIPS",
         pad_y = FALSE,
         fun_dist = extract_at,
         x = ncelev,
@@ -734,14 +852,15 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
           regions = nccnty,
           .debug = TRUE,
           pad_y = TRUE,
-          regions_id = "GEOID",
+          regions_id = "FIPS",
           fun_dist = nearest,
           x = ncsamp,
-          y = nctrct
+          y = terra::vect(nccntygrid)
         )
     )
   )
   future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -749,6 +868,7 @@ testthat::test_that("par_hierarchy: define level by substring", {
   withr::local_package("terra")
   withr::local_package("sf")
   withr::local_package("future")
+  withr::local_package("mirai")
   withr::local_package("future.apply")
   withr::local_package("future.mirai")
   withr::local_package("dplyr")
@@ -763,10 +883,30 @@ testthat::test_that("par_hierarchy: define level by substring", {
   )
   withr::local_seed(202407)
 
-  ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-  nccnty <- sf::st_read(ncpath, layer = "county")
-  nctrct <- sf::st_read(ncpath, layer = "tracts")
-  ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
+  nccnty <- sf::st_read(
+    system.file("shape/nc.shp", package = "sf")
+  )
+  nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+  nccntygrid <- sf::st_make_grid(nccnty, n = c(200, 100))
+  nccntygrid <- sf::st_as_sf(nccntygrid)
+  nccntygrid$GEOID <- sprintf("%05d", seq_len(nrow(nccntygrid)))
+  suppressWarnings(
+    nccntygrid <- sf::st_intersection(nccntygrid, nccnty)
+  )
+
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  ncpnts <- terra::vect(ncp)
+
+  ## Resampled SRTM data in NC
+  ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+  terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+  # Using raster path
+  ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+  terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
   ncelev <- terra::rast(ncelevpath)
 
   ncsamp <-
@@ -783,12 +923,12 @@ testthat::test_that("par_hierarchy: define level by substring", {
   testthat::expect_no_error(
     residc <-
       par_hierarchy(
-        regions = nctrct,
-        regions_id = "GEOID",
+        regions = nccntygrid,
+        regions_id = "FIPS",
         length_left = 5L,
         .debug = TRUE,
         fun_dist = extract_at,
-        y = nctrct,
+        y = nccntygrid,
         x = ncelev,
         id = "GEOID",
         func = "mean"
@@ -797,22 +937,24 @@ testthat::test_that("par_hierarchy: define level by substring", {
   testthat::expect_true(is.data.frame(residc))
 
   # bare integers with different lengths: warning message
-  nctrct$qid <- seq_len(nrow(nctrct))
+  nccntygrid$qid <- seq_len(nrow(nccntygrid))
   testthat::expect_message(
     residc <-
       par_hierarchy(
-        regions = nctrct,
+        regions = nccntygrid,
         regions_id = "qid",
         length_left = 2L,
         .debug = TRUE,
         fun_dist = extract_at,
-        y = nctrct,
+        y = nccntygrid,
         x = ncelev,
         id = "GEOID",
         func = "mean"
       )
   )
 
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -833,12 +975,13 @@ testthat::test_that("generic function should be parallelized properly", {
   )
 
   # main test
-  pnts <- readRDS(
-    system.file("extdata/nc_random_point.rds", package = "chopin")
-  )
-  pnts <- terra::vect(pnts)
-  rd1 <-
-    terra::vect(system.file("extdata/ncroads_first.gpkg", package = "chopin"))
+  ## Generated random points in NC
+  data("ncpoints", package = "chopin")
+  ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+  ncp$pid <- seq_len(nrow(ncp))
+  pnts <- terra::vect(ncp)
+
+  rd1 <- terra::vect(roadpath)
 
   pnts <- terra::project(pnts, "EPSG:5070")
   rd1 <- terra::project(rd1, "EPSG:5070")
@@ -867,6 +1010,8 @@ testthat::test_that("generic function should be parallelized properly", {
     "terra inputs detected in both x and y. Please replace x and y to file paths to proceed."
   )
 
+  future::plan(future::sequential)
+  mirai::daemons(0)
 })
 
 
@@ -887,12 +1032,30 @@ testthat::test_that(
       )
     )
     future::plan(future.mirai::mirai_multisession, workers = 2L)
-    ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-    nccnty <- sf::st_read(ncpath, layer = "county")
-    nccnty <- terra::vect(nccnty)
-    ncelev <-
-      system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    ncelev <- terra::rast(ncelev)
+
+
+    nccnty <- sf::st_read(
+      system.file("shape/nc.shp", package = "sf")
+    )
+    nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+    ncpath <- file.path(tempdir(check = TRUE), "nccnty.gpkg")
+    sf::st_write(nccnty, ncpath, "nccnty", delete_dsn = TRUE)
+
+    ## Generated random points in NC
+    data("ncpoints", package = "chopin")
+    ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+    ncp$pid <- seq_len(nrow(ncp))
+    ncpnts <- terra::vect(ncp)
+
+    ## Resampled SRTM data in NC
+    ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+    terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+    # Using raster path
+    ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+    terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
+    ncelev <- terra::rast(ncelevpath)
+
     tdir <- tempdir(check = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test1.tif"), overwrite = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test2.tif"), overwrite = TRUE)
@@ -919,6 +1082,7 @@ testthat::test_that(
     testthat::expect_true(!anyNA(res))
 
     future::plan(future::sequential)
+    mirai::daemons(0)
   }
 )
 
@@ -942,11 +1106,29 @@ testthat::test_that(
       )
     )
     future::plan(future::multicore, workers = 2L)
-    ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-    nccnty <- terra::vect(ncpath, layer = "county")
-    ncelev <-
-      system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    ncelev <- terra::rast(ncelev)
+
+    nccnty <- sf::st_read(
+      system.file("shape/nc.shp", package = "sf")
+    )
+    nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+    ncpath <- file.path(tempdir(check = TRUE), "nccnty.gpkg")
+    sf::st_write(nccnty, ncpath, "nccnty", delete_dsn = TRUE)
+
+    ## Generated random points in NC
+    data("ncpoints", package = "chopin")
+    ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+    ncp$pid <- seq_len(nrow(ncp))
+    ncpnts <- terra::vect(ncp)
+
+    ## Resampled SRTM data in NC
+    ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+    terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+    # Using raster path
+    ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+    terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
+    ncelev <- terra::rast(ncelevpath)
+
     tdir <- tempdir(check = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test1.tif"), overwrite = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test2.tif"), overwrite = TRUE)
@@ -967,11 +1149,11 @@ testthat::test_that(
         )
       )
     )
-    future::plan(future::sequential)
 
+    future::plan(future::sequential)
+    mirai::daemons(0)
   }
 )
-
 
 
 testthat::test_that(
@@ -993,11 +1175,29 @@ testthat::test_that(
       )
     )
     future::plan(future.mirai::mirai_multisession, workers = 2L)
-    ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-    nccnty <- sprintf("GPKG:%s:%s", ncpath, "county")
-    ncelev <-
-      system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    ncelev <- terra::rast(ncelev)
+
+    nccnty <- sf::st_read(
+      system.file("shape/nc.shp", package = "sf")
+    )
+    nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+    ncpath <- file.path(tempdir(check = TRUE), "nccnty.gpkg")
+    sf::st_write(nccnty, ncpath, "nccnty", delete_dsn = TRUE)
+
+    ## Generated random points in NC
+    data("ncpoints", package = "chopin")
+    ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+    ncp$pid <- seq_len(nrow(ncp))
+    ncpnts <- terra::vect(ncp)
+
+    ## Resampled SRTM data in NC
+    ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+    terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+    # Using raster path
+    ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+    terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
+    ncelev <- terra::rast(ncelevpath)
+
     tdir <- tempdir(check = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test1.tif"), overwrite = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test2.tif"), overwrite = TRUE)
@@ -1032,21 +1232,22 @@ testthat::test_that(
     # error case: function loading with ::
     testthat::expect_no_error(
       suppressWarnings(
-      nut <- par_multirasters(
-        filenames = testfiles_corrupted,
-        .debug = TRUE,
-        fun_dist = terra::extract,
-        y = nccnty,
-        x = ncelev,
-        ID = TRUE,
-        fun = mean
-      )
+        nut <- par_multirasters(
+          filenames = testfiles_corrupted,
+          .debug = TRUE,
+          fun_dist = terra::extract,
+          y = nccnty,
+          x = ncelev,
+          ID = TRUE,
+          fun = mean
+        )
       )
     )
 
     testthat::expect_s3_class(nut, "data.frame")
-    future::plan(future::sequential)
 
+    future::plan(future::sequential)
+    mirai::daemons(0)
   }
 )
 
@@ -1070,12 +1271,30 @@ testthat::test_that(
       )
     )
     future::plan(future.mirai::mirai_multisession, workers = 2L)
-    ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
-    nccnty <- sprintf("GPKG:%s:%s", ncpath, "county")
-    suppressWarnings(nccnty <- sf::st_read(nccnty))
-    ncelev <-
-      system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    ncelev <- terra::rast(ncelev)
+
+
+    nccnty <- sf::st_read(
+      system.file("shape/nc.shp", package = "sf")
+    )
+    nccnty <- sf::st_transform(nccnty, "EPSG:5070")
+    ncpath <- file.path(tempdir(check = TRUE), "nccnty.gpkg")
+    sf::st_write(nccnty, ncpath, "nccnty", delete_dsn = TRUE)
+
+    ## Generated random points in NC
+    data("ncpoints", package = "chopin")
+    ncp <- sf::st_as_sf(ncpoints, coords = c("X", "Y"), crs = "EPSG:5070")
+    ncp$pid <- seq_len(nrow(ncp))
+    ncpnts <- terra::vect(ncp)
+
+    ## Resampled SRTM data in NC
+    ras <- terra::rast(nccnty, nrow = 1000, ncol = 2200)
+    terra::values(ras) <- rgamma(2.2e6, 4, 2)
+
+    # Using raster path
+    ncelevpath <- file.path(tempdir(check = TRUE), "ncelev.tif")
+    terra::writeRaster(ras, ncelevpath, overwrite = TRUE)
+    ncelev <- terra::rast(ncelevpath)
+
     tdir <- tempdir(check = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test1.tif"), overwrite = TRUE)
     terra::writeRaster(ncelev, file.path(tdir, "test2.tif"), overwrite = TRUE)
@@ -1117,6 +1336,7 @@ testthat::test_that(
     testthat::expect_true(sum(!is.na(nut$error_message)) == 1L)
 
     future::plan(future::sequential)
+    mirai::daemons(0)
   }
 )